styling changes. also included concurrent updating of task positions

+++ Kanban/db/migrate/20181030233355_create_tasks.rb
@@ -4,8 +4,8 @@ class CreateTasks < ActiveRecord::Migration[5.2]
       t.string :name
       t.text :description
       t.integer :progress
-      t.integer :x
-      t.integer :y
+      t.float :x
+      t.float :y
       t.references :board, foreign_key: true
       t.timestamps
     end
